+++ server/gaas/text_gen/operations/json.py
@@ -1,16 +1,24 @@
 import logging
-from typing import Any, Dict, Optional
+from typing import Optional
 import guidance
 import json
-from gaas.text_gen.abstract_text_gen import AbstractTextGen
 import torch
+from gaas.model_manager.model_manager import ModelManager
 
 logger = logging.getLogger(__name__)
 
 
-class JSON(AbstractTextGen):
-    def __init__(self, model_name: str, **kwargs):
-        super().__init__(model_name=model_name, **kwargs)
+class JSON:
+    """Example of a text generation class using the shared model manager."""
+
+    def __init__(self, model_manager: ModelManager, **kwargs):
+        self.model_manager = model_manager
+        self.model = model_manager.model
+        self.tokenizer = model_manager.tokenizer
+        self.pipe = model_manager.pipe
+        self.device = model_manager.device
+
+        # JSON-specific initialization
         self.model.to(torch.float16)
         self.guidance_model = guidance.models.Transformers(
             model=self.model, tokenizer=self.tokenizer, echo=False, dtype=torch.float16
@@ -18,24 +26,9 @@ class JSON(AbstractTextGen):
 
     def generate(
         self, prompt: str, json_schema: str, context: Optional[str] = None, **kwargs
-    ) -> Dict[str, Any]:
-        """
-        Convert text data to JSON format based on a provided schema.
-
-        Args:
-            text_data: The input text to convert to JSON
-            json_schema: A JSON schema string defining the expected output structure
-            context: Custom context/prompt to guide the JSON conversion
-
-        Returns:
-            Dict containing:
-                - input: Original text data
-                - schema: Original schema
-                - output: Generated JSON or error message
-        """
+    ):
         response = {"input": prompt, "schema": json_schema, "output": None}
 
-        # Parse and validate JSON schema
         try:
             parsed_schema = json.loads(json_schema)
             if not isinstance(parsed_schema, dict):
@@ -48,7 +41,6 @@ class JSON(AbstractTextGen):
             return response
 
         try:
-            # Set up conversion context
             default_context = (
                 "You are an expert who converts English text to JSON. "
                 "When data is not available, use -111 as default value. "
@@ -57,13 +49,11 @@ class JSON(AbstractTextGen):
             conversion_prompt = f"{context or default_context}\n{prompt}"
 
             with torch.amp.autocast("cuda", dtype=torch.float16):
-                # Generate JSON using guidance
                 json_output = self.guidance_model + conversion_prompt
                 json_output += guidance.json(
                     name="generated_object", schema=parsed_schema
                 )
 
-            # Extract and validate generated JSON
             generated_json = json_output.get("generated_object")
             if not generated_json:
                 raise ValueError("No JSON output generated")
